Fix all build errors for GCC 13 / C++23 on Ubuntu 24.04
CMakeLists.txt (top-level):
  - Add C to LANGUAGES (was CXX-only) so pure-C targets like netlink_test
    get a valid linker language.
  - Lower standard from C++26 to C++23; GCC 13 supports C++23 fully.
  - Guard C++-only warning flags (-Wnon-virtual-dtor, -Wold-style-cast,
    -Woverloaded-virtual) with $<COMPILE_LANGUAGE:CXX> so they are never
    applied to C translation units.

common/include/print  (NEW):
  - Compatibility shim for <print>.  GCC 13's libstdc++ ships C++23
    std::format but not std::print / std::println (<print> header only
    arrived in GCC 14).  Provides print/println overloads for stdout,
    FILE* and std::ostream& (the C++26 P3107 overloads used by
    std::println(std::cerr, ...) throughout the codebase).
  - Placed in common/include so it is propagated automatically to srmd
    and sra via common_lib's PUBLIC include path.

client/include/client/route_client.hpp:
  - makeContext() return type changed from grpc::ClientContext (private
    copy constructor in gRPC 1.51, making return-by-value ill-formed) to
    std::unique_ptr<grpc::ClientContext>.

client/src/route_client.cpp:
  - Updated makeContext() implementation to return unique_ptr.
  - All call sites changed from &ctx to ctx.get().
  - resp.code() (srmd::v1::StatusCode, a protobuf enum) replaced by
    StatusCode_Name(resp.code()) in std::format calls; GCC 13 has no
    built-in formatter for unscoped enums.
  - Added #include <memory>.

client/src/sot_config.cpp:
  - Added std::formatter<boost::core::basic_string_view<char>>
    specialisation so boost::json object keys can be passed directly to
    std::format (GCC 13 has no such specialisation; GCC 14 does).

server/src/main.cpp:
  - Added missing #include <fstream> (std::ifstream was an incomplete
    type in the TLS credential loader lambda).
  - daemon.shouldReload() result was discarded (nodiscard); now handled
    with an explicit if-branch that logs the reload signal.

server/src/netlink.c:
  - attrlen changed from int to unsigned int to silence -Wconversion /
    -Wsign-conversion warnings from RTM_PAYLOAD / RTA_NEXT macros.

server/src/netlink_test.c:
  - Millisecond field in timestamp() changed from %03ld (long) to %03d
    (int) with explicit cast to fix -Wformat-truncation warning.

server/CMakeLists.txt / client/CMakeLists.txt:
  - CreateVersionedBinary.cmake post-build commands corrected:
    BUILD_DIR was CMAKE_CURRENT_BINARY_DIR (build/server or build/client)
    but IncrementBuildNumber writes build_number.txt to the root build
    dir; changed to CMAKE_BINARY_DIR.

+++ server/src/netlink.c
@@ -113,7 +113,7 @@ static void nl_dispatch(const struct nlmsghdr *nlh,
     route.protocol = rtm->rtm_protocol;
     route.table    = rtm->rtm_table; /* May be overridden by RTA_TABLE below */
 
-    int attrlen = RTM_PAYLOAD(nlh);
+    unsigned int attrlen = (unsigned int)RTM_PAYLOAD(nlh);
     for (const struct rtattr *rta = RTM_RTA(rtm);
          RTA_OK(rta, attrlen);
          rta = RTA_NEXT(rta, attrlen))
